src/main/java/com/kirayim/homesec/webapp/controller/CameraController.java
package com.kirayim.homesec.webapp.controller;

import org.appfuse.dao.SearchException;
import org.appfuse.service.GenericManager;
import com.kirayim.homesec.model.Camera;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Controller;
import org.springframework.ui.ExtendedModelMap;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

@Controller
@RequestMapping("/cameras*")
public class CameraController {
    private GenericManager<Camera, Long> cameraManager;

    @Autowired
    public void setCameraManager(@Qualifier("cameraManager") GenericManager<Camera, Long> cameraManager) {
        this.cameraManager = cameraManager;
    }

    @RequestMapping(method = RequestMethod.GET)
    public Model handleRequest(@RequestParam(required = false, value = "q") String query)
    throws Exception {
        Model model = new ExtendedModelMap();
        try {
            model.addAttribute(cameraManager.search(query, Camera.class));
        } catch (SearchException se) {
            model.addAttribute("searchError", se.getMessage());
            model.addAttribute(cameraManager.getAll());
        }
        return model;
    }
}
